Add demo on read_csv() add chunk to fix printed wd
we want the printed wd error to be from the root, so as not to confuse participants. Because we’re using Rmd files, when the chunk code is printed, the file paths are relative to the Rmd file, but we want participants to try these examples relative to the root dir of the project.

Signed-off-by: egouldo <elisebgould@gmail.com>